Pass all task arguments not just data ids
The example I start with required very few arguments to run the task.
Thus I decided to make Executor construct the actual command line
for the it.  However, even a relatively simple pipeline can have complex
command line structure.  Thus I modified Executor to pass task arguments
as they are with an exception arguments it needs to be aware of, i.e.
locations of the input and output dataset repositories.