Do not allow for silent removal of existing repos
Executor can either use existing input repository or create a new one
from scratch in a given location.  The mode of operation is selected
automatically based on presence certain fields in the job description.
Currently, when creating a new dataset repository Executor wipes out any
pre-existing content at the specified location.  This behavior is not
safe as it might lead to accidental removal of an existing dataset
repository if a malformed job description would be used.

I modified Executor to treat the path to the input dataset repository as
read only location by default.  If a user wishes Executor to create a
dataset repository from scratch at a given location, the location must
be empty and explicitly flagged as writable in a job description.